Ajustes

+++ src/main/java/com/hexacta/altapersonas/exception/PersonaNotFoundException.java
@@ -1,7 +1,7 @@
 package com.hexacta.altapersonas.exception;
 
 public class PersonaNotFoundException extends RuntimeException{
-    public PersonaNotFoundException(int id){
+    public PersonaNotFoundException(Long id){
         super("No se pudo encontrar el usuario con el id "+id);
     }
 }
